+++ tests/test_query.py
@@ -126,24 +126,7 @@ def test_in_query(es_data, engine):
 
 
 def test_select_muiltipaths(es_data, engine):
-    """
-    pp res["_source"]
-{'organization_resource': {'address': [{'city': 'Den Burg;Jeg fik bøde på '
-                                                '5.000 kroner i et andet '
-                                                'orkester. først søge '
-                                                'permanent ophold i år 2032',
-                                        'country': 'NLD',
-                                        'line': ['Galapagosweg 91'],
-                                        'postalCode': '9105 PZ',
-                                        'use': 'work'},
-                                       {'city': 'Den Burg',
-                                        'country': 'NLD',
-                                        'line': ['PO Box 2311'],
-                                        'postalCode': '9100 AA',
-                                        'use': 'work'}],
-                           'name': 'Burgers University Medical Center'}}
-(Pdb) c
-    """
+    """ """
     builder = Q_(resource="Organization", engine=engine).select(
         "Organization.name", "Organization.address"
     )
@@ -151,3 +134,24 @@ def test_select_muiltipaths(es_data, engine):
     result = builder(async_result=False).fetchall()
 
     assert len(result.body[0]) == 2
+
+
+def test_result_count(es_data, engine):
+    """ """
+    conn, meta_info = es_data
+    load_organizations_data(conn, 5)
+    builder = Q_(resource="Organization", engine=engine)
+    builder = builder.where(T_("Organization.active") == V_("true"))
+    total = builder(async_result=False).count()
+
+    assert total == 6
+
+
+def test_result_empty(es_data, engine):
+    """ """
+    conn, meta_info = es_data
+    load_organizations_data(conn, 5)
+    builder = Q_(resource="Organization", engine=engine)
+    builder = builder.where(T_("Organization.active") == V_("false"))
+    empty = builder(async_result=False).empty()
+    assert empty is True