+++ HISTORY.rst
@@ -9,6 +9,7 @@ Improvements
 
 - Now full select features are accepted, mean you can provide multiple path in select section. for example ``select(Patient.name, Patient.gender)``.
 
+- FHIRPath ``count()`` and ``empty()`` functions are supported.
 
 Breakings